+++ appEditors/appGCodeEditor.py
@@ -197,7 +197,7 @@ class AppGCodeEditor(QtCore.QObject):
             job_item = QtWidgets.QTableWidgetItem(job_item_txt)
 
             # -------------------- TOOL SHAPE ------------------------------------- #
-            tool_type_item_options = ["C1", "C2", "C3", "C4", "B", "V"]
+            tool_type_item_options = ["C1", "C2", "C3", "C4", "B", "V", "L"]
             try:
                 tool_shape_item_txt = tool_type_item_options[dia_value['data']['tools_mill_tool_shape']]
             except TypeError: