+++ appObjects/FlatCAMGeometry.py
@@ -123,7 +123,7 @@ class GeometryObject(FlatCAMObj, Geometry):
 
         self.offset_item_options = ["Path", "In", "Out", "Custom"]
         self.job_item_options = [_('Roughing'), _('Finishing'), _('Isolation'), _('Polishing')]
-        self.tool_type_item_options = ["C1", "C2", "C3", "C4", "B", "V"]
+        self.tool_type_item_options = ["C1", "C2", "C3", "C4", "B", "V", "L"]
 
         # flag to store if the V-Shape tool is selected in self.ui.geo_tools_table
         self.v_tool_type = None